refactor(src/layouts): JavaScript → TypeScript
